+++ .idea/workspace.xml
@@ -4,25 +4,13 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="5933f241-2562-4785-ac50-34f4229d0992" name="Changes" comment="feat: 更改标题">
-      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/config/Config.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/config/Param.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/config/Parameters.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/LogETLTask.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/api-common/src/main/java/com/api/common/entity/ReportLog.java" beforeDir="false" afterPath="$PROJECT_DIR$/api-common/src/main/java/com/api/common/entity/ReportLog.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/api-common/src/main/java/com/api/common/entity/ReportLogPv.java" beforeDir="false" afterPath="$PROJECT_DIR$/api-common/src/main/java/com/api/common/entity/ReportLogPv.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/pom.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/LogEtlMainTask.java" beforeDir="false" />
+    <list default="true" id="5933f241-2562-4785-ac50-34f4229d0992" name="Changes" comment="feat: 初始化pv数据sink到mysql">
+      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/sink/LogPvMysqlSink.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/TaskRunner.java" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/window/LogPvWatermarkAssigner.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/LogETLTask.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/LogEtlTask.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/ReportDataLoadTask.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/task/ReportDataLoadTask.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/MysqlUtils.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/MysqlUtils.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/PropertyUtils.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/PropertyUtils.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/RedisUtils.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/util/RedisUtils.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/window/ReportPvWindow.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/window/ReportPvWindow.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/resources/config.properties" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/resources/config.properties" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flink-job/src/main/resources/log4j.properties" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/resources/log4j.properties" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/pom.xml" beforeDir="false" afterPath="$PROJECT_DIR$/pom.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/window/ReportPvWindow.java" beforeDir="false" afterPath="$PROJECT_DIR$/flink-job/src/main/java/com/flink/job/window/LogPvWindow.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -121,9 +109,22 @@
       <recent name="com.flink.job.config" />
     </key>
   </component>
-  <component name="RunManager" selected="Application.LogETLTask">
-    <configuration name="LogETLTask" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="com.flink.job.task.LogETLTask" />
+  <component name="RunManager" selected="Application.EtlTaskRunner">
+    <configuration name="EtlTaskRunner" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="com.flink.job.task.EtlTaskRunner" />
+      <module name="flink-job" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="com.flink.job.task.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration name="LogEtlTask" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="com.flink.job.task.LogEtlTask" />
       <module name="flink-job" />
       <extension name="coverage">
         <pattern>
@@ -157,8 +158,9 @@
     </configuration>
     <recent_temporary>
       <list>
-        <item itemvalue="Application.LogETLTask" />
+        <item itemvalue="Application.EtlTaskRunner" />
         <item itemvalue="Application.ReportDataLoadTask" />
+        <item itemvalue="Application.LogEtlTask" />
       </list>
     </recent_temporary>
   </component>
@@ -171,7 +173,7 @@
       <option name="presentableId" value="Default" />
       <updated>1692443282937</updated>
       <workItem from="1692443284190" duration="21370000" />
-      <workItem from="1692508912295" duration="2964000" />
+      <workItem from="1692508912295" duration="3455000" />
     </task>
     <task id="LOCAL-00001" summary="feat: 初始化">
       <created>1692443338309</created>
@@ -243,7 +245,14 @@
       <option name="project" value="LOCAL" />
       <updated>1692456203978</updated>
     </task>
-    <option name="localTasksCounter" value="11" />
+    <task id="LOCAL-00011" summary="feat: 初始化pv数据sink到mysql">
+      <created>1692512064214</created>
+      <option name="number" value="00011" />
+      <option name="presentableId" value="LOCAL-00011" />
+      <option name="project" value="LOCAL" />
+      <updated>1692512064214</updated>
+    </task>
+    <option name="localTasksCounter" value="12" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -269,6 +278,7 @@
     <MESSAGE value="feat: 初始化flink模块" />
     <MESSAGE value="feat: 初始化flink的读取任务" />
     <MESSAGE value="feat: 更改标题" />
-    <option name="LAST_COMMIT_MESSAGE" value="feat: 更改标题" />
+    <MESSAGE value="feat: 初始化pv数据sink到mysql" />
+    <option name="LAST_COMMIT_MESSAGE" value="feat: 初始化pv数据sink到mysql" />
   </component>
 </project>
